tag_creator/requirements.txt
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.11
decorator==5.1.1
idna==3.3
ipython==7.31.1
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.21.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pycairo==1.19.0
Pygments==2.11.2
qrcode==7.3.1
requests==2.27.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
